refactor(web): extract pure chat rendering components from ChatCenter
Move MessageBubble, ToolRecordDetails, ErrorBlock, ChannelBadge and
friends into react/chat/MessageBubble.tsx as exported, props-driven
components with no store/api imports, so apps/showcase can replay
recorded sessions with the exact same rendering code. No behavior
change.
